# implement in-place decode