--- a/decoder_test.go
+++ b/decoder_test.go
@@ -152,11 +152,11 @@ func BenchmarkDecode16_10With1Error(b *testing.B) {
 	b.StopTimer()
 	d := NewDecoder(QRCodeField256)
 	complete := getcomplete()
-	complete[1] = complete[1] + 1
 	b.SetBytes(int64(len(QRCodeTestData) * b.N))
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
-		if n, err := d.Decode(makecopy(complete), len(QRCodeTestECC)); err != nil || n != 1 {
+		complete[1] = complete[1] + 1
+		if n, err := d.Decode(complete, len(QRCodeTestECC)); err != nil || n != 1 {
 			b.Fail()
 		}
 	}
@@ -186,12 +186,12 @@ func BenchmarkDecode128_16With1Error(b *testing.B) {
 	e := NewEncoder(QRCodeField256, len(ecc))
 	e.Encode(data, ecc)
 	d := NewDecoder(QRCodeField256)
-	data[1] = data[1] + 1
 	complete := append(data, ecc...)
 	b.SetBytes(int64(len(data) * b.N))
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
-		if n, err := d.Decode(makecopy(complete), len(ecc)); err != nil || n != 1 {
+		complete[1] = complete[1] + 1
+		if n, err := d.Decode(complete, len(ecc)); err != nil || n != 1 {
 			b.Fail()
 		}
 	}
@@ -204,13 +204,13 @@ func BenchmarkDecode128_16With2Error(b *testing.B) {
 	e := NewEncoder(QRCodeField256, len(ecc))
 	e.Encode(data, ecc)
 	d := NewDecoder(QRCodeField256)
-	data[1] = data[1] + 1
-	ecc[1] = ecc[1] + 1
 	complete := append(data, ecc...)
 	b.SetBytes(int64(len(data) * b.N))
 	b.StartTimer()
 	for i := 0; i < b.N; i++ {
-		if n, err := d.Decode(makecopy(complete), len(ecc)); err != nil || n != 2 {
+		complete[1] = complete[1] + 1
+		complete[len(data)+1] = complete[len(data)+1] + 1
+		if n, err := d.Decode(complete, len(ecc)); err != nil || n != 2 {
 			b.Fail()
 		}
 	}