--- a/example_test.go
+++ b/example_test.go
@@ -7,19 +7,17 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 or implied. See the License for the specific language governing permissions and
 limitations under the License. */
 
-package rs_test
+package rs
 
 import (
 	"fmt"
-
-	"github.com/maruel/rs"
 )
 
 func ExampleNewEncoder() {
 	data := []byte("hello, world")
 	fmt.Printf("Original data: %s\n", data)
 	ecc := make([]byte, 2)
-	e := rs.NewEncoder(rs.QRCodeField256, len(ecc))
+	e := NewEncoder(QRCodeField256, len(ecc))
 	e.Encode(data, ecc)
 	fmt.Printf("ECC bytes: %v\n", ecc)
 	// Output:
@@ -31,11 +29,25 @@ func ExampleNewDecoder() {
 	data := []byte("hello, wXrld")
 	ecc := []byte{171, 167}
 	fmt.Printf("Corrupted data: %s\n", data)
-	d := rs.NewDecoder(rs.QRCodeField256)
+	d := NewDecoder(QRCodeField256)
 	if nb, err := d.Decode(data, ecc); err != nil || nb != 1 {
 		fmt.Printf("Expected 1 fix, for %d. Error: %s\n", nb, err)
 	}
 	fmt.Printf("Fixed data: %s\n", data)
+	// Output:
+	// Corrupted data: hello, wXrld
+	// Fixed data: hello, world
+}
+
+func ExampleDecoder_DecodeInPlace() {
+	data := []byte("hello, wXrld\xAB\xA7")
+	eccLen := 2
+	fmt.Printf("Corrupted data: %s\n", data[:len(data)-eccLen])
+	d := NewDecoder(QRCodeField256)
+	if nb, err := d.DecodeInPlace(data, eccLen); err != nil || nb != 1 {
+		fmt.Printf("Expected 1 fix, for %d. Error: %s\n", nb, err)
+	}
+	fmt.Printf("Fixed data: %s\n", data[:len(data)-eccLen])
 	// Output:
 	// Corrupted data: hello, wXrld
 	// Fixed data: hello, world