--- a/example_test.go
+++ b/example_test.go
@@ -7,17 +7,19 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 or implied. See the License for the specific language governing permissions and
 limitations under the License. */
 
-package rs
+package rs_test
 
 import (
 	"fmt"
+
+	"github.com/maruel/rs"
 )
 
 func ExampleNewEncoder() {
 	data := []byte("hello, world")
 	fmt.Printf("Original data: %s\n", data)
 	ecc := make([]byte, 2)
-	e := NewEncoder(QRCodeField256, len(ecc))
+	e := rs.NewEncoder(rs.QRCodeField256, len(ecc))
 	e.Encode(data, ecc)
 	fmt.Printf("ECC bytes: %v\n", ecc)
 	// Output:
@@ -29,7 +31,7 @@ func ExampleNewDecoder() {
 	data := []byte("hello, wXrld")
 	ecc := []byte{171, 167}
 	fmt.Printf("Corrupted data: %s\n", data)
-	d := NewDecoder(QRCodeField256)
+	d := rs.NewDecoder(rs.QRCodeField256)
 	if nb, err := d.Decode(data, ecc); err != nil || nb != 1 {
 		fmt.Printf("Expected 1 fix, for %d. Error: %s\n", nb, err)
 	}
@@ -43,7 +45,7 @@ func ExampleDecoder_DecodeInPlace() {
 	data := []byte("hello, wXrld\xAB\xA7")
 	eccLen := 2
 	fmt.Printf("Corrupted data: %s\n", data[:len(data)-eccLen])
-	d := NewDecoder(QRCodeField256)
+	d := rs.NewDecoder(rs.QRCodeField256)
 	if nb, err := d.DecodeInPlace(data, eccLen); err != nil || nb != 1 {
 		fmt.Printf("Expected 1 fix, for %d. Error: %s\n", nb, err)
 	}