--- a/example_test.go
+++ b/example_test.go
@@ -28,25 +28,11 @@ func ExampleNewEncoder() {
 }
 
 func ExampleNewDecoder() {
-	data := []byte("hello, wXrld")
-	ecc := []byte{171, 167}
-	fmt.Printf("Corrupted data: %s\n", data)
-	d := rs.NewDecoder(rs.QRCodeField256)
-	if nb, err := d.Decode(data, ecc); err != nil || nb != 1 {
-		fmt.Printf("Expected 1 fix, for %d. Error: %s\n", nb, err)
-	}
-	fmt.Printf("Fixed data: %s\n", data)
-	// Output:
-	// Corrupted data: hello, wXrld
-	// Fixed data: hello, world
-}
-
-func ExampleDecoder_DecodeInPlace() {
 	data := []byte("hello, wXrld\xAB\xA7")
 	eccLen := 2
 	fmt.Printf("Corrupted data: %s\n", data[:len(data)-eccLen])
 	d := rs.NewDecoder(rs.QRCodeField256)
-	if nb, err := d.DecodeInPlace(data, eccLen); err != nil || nb != 1 {
+	if nb, err := d.Decode(data, eccLen); err != nil || nb != 1 {
 		fmt.Printf("Expected 1 fix, for %d. Error: %s\n", nb, err)
 	}
 	fmt.Printf("Fixed data: %s\n", data[:len(data)-eccLen])