I implement the in-place decode and add a test for it.