@karaketir16 2024-05-12
Here Benchmark results, but I am not sure I benchmarked correctly and I don't know how to analyze this results.

Old Decode
```
goos: linux
goarch: amd64
pkg: github.com/maruel/rs
cpu: 12th Gen Intel(R) Core(TM) i5-12450H
BenchmarkDecode16_10-12               	 2445758	       473.5 ns/op	82648331.65 MB/s	      48 B/op	       2 allocs/op
BenchmarkDecode16_10With1Error-12     	 1256457	       949.2 ns/op	21180130.20 MB/s	     528 B/op	      28 allocs/op
BenchmarkDecode128_16-12              	  198171	      6036 ns/op	4202141.26 MB/s	     160 B/op	       2 allocs/op
BenchmarkDecode128_16With1Error-12    	  181016	      6586 ns/op	3518293.74 MB/s	     790 B/op	      28 allocs/op
BenchmarkDecode128_16With2Error-12    	  137966	      7320 ns/op	2412675.47 MB/s	    1176 B/op	      47 allocs/op
PASS
ok  	github.com/maruel/rs	7.444s
```
New Decode
```
goos: linux
goarch: amd64
pkg: github.com/maruel/rs
cpu: 12th Gen Intel(R) Core(TM) i5-12450H
BenchmarkDecode16_10-12               	 2521464	       477.9 ns/op	84420616.63 MB/s	      16 B/op	       1 allocs/op
BenchmarkDecode16_10With1Error-12     	 1000000	      1049 ns/op	15254901.10 MB/s	     512 B/op	      27 allocs/op
BenchmarkDecode128_16-12              	  184364	      6157 ns/op	3832831.58 MB/s	      16 B/op	       1 allocs/op
BenchmarkDecode128_16With1Error-12    	  178500	      6763 ns/op	3378207.00 MB/s	     662 B/op	      27 allocs/op
BenchmarkDecode128_16With2Error-12    	  138470	      7644 ns/op	2318786.01 MB/s	    1032 B/op	      45 allocs/op
PASS
ok  	github.com/maruel/rs	6.387s
```