@karaketir16 2024-05-12
I updated the benchmark with `makecopy` removed. I also updated the benchmark for the old decode function while running.

### Old Decode
```
goos: linux
goarch: amd64
pkg: github.com/maruel/rs
cpu: 12th Gen Intel(R) Core(TM) i5-12450H
BenchmarkDecode16_10-12               	 2407120	       498.6 ns/op	77244958.62 MB/s	      48 B/op	       2 allocs/op
BenchmarkDecode16_10With1Error-12     	 1305115	       941.2 ns/op	22186455.84 MB/s	     512 B/op	      27 allocs/op
BenchmarkDecode128_16-12              	  196250	      6122 ns/op	4103061.96 MB/s	     160 B/op	       2 allocs/op
BenchmarkDecode128_16With1Error-12    	  181203	      6679 ns/op	3472787.80 MB/s	     662 B/op	      27 allocs/op
BenchmarkDecode128_16With2Error-12    	  164697	      7258 ns/op	2904691.98 MB/s	    1032 B/op	      45 allocs/op
PASS
ok  	github.com/maruel/rs	7.689s

```

### New Decode
```
goos: linux
goarch: amd64
pkg: github.com/maruel/rs
cpu: 12th Gen Intel(R) Core(TM) i5-12450H
BenchmarkDecode16_10-12               	 2505919	       465.6 ns/op	86111209.87 MB/s	      16 B/op	       1 allocs/op
BenchmarkDecode16_10With1Error-12     	 1312581	       899.8 ns/op	23340185.02 MB/s	     480 B/op	      26 allocs/op
BenchmarkDecode128_16-12              	  202328	      6011 ns/op	4308775.31 MB/s	      16 B/op	       1 allocs/op
BenchmarkDecode128_16With1Error-12    	  179751	      6567 ns/op	3503619.46 MB/s	     518 B/op	      26 allocs/op
BenchmarkDecode128_16With2Error-12    	  153178	      7212 ns/op	2718774.60 MB/s	     888 B/op	      44 allocs/op
PASS
ok  	github.com/maruel/rs	7.481s
```